Fixes semicolon parsing within a url
The dynamic block is pretty hookey, but I think it might work. It is
doing a manual lookbehind to see if there is an open parenthesis. Not
the prettiest code, but it should work.

It is also downcasing the base64 encoded data, which could cause
collision issues, but encoded images are so uncommon that I don't think
it will be a big issue. For now.

refs: #46