Adds an experimental flag to ignore sass mixin declarations
Users can now do this with --ignore-sass-mixins.

This is a heavily requested feature. Some users don't care about
duplication coming from sass mixins. So this will inject a special
comment before/after the properties from the mixin and the parser will
ignore it.

This will mess with reporting selector line numbers if I implement that
in the future.

I'm not sold on this being the default yet.

refs: #25, #59